+++ sgl-kernel/csrc/torch_extension_rocm.cc
@@ -56,21 +56,21 @@ TORCH_LIBRARY_EXPAND(sgl_kernel, m) {
 
   // quick allreduce
 #ifdef USE_ROCM
-  custom_ar.def(
+  m.def(
       "qr_all_reduce(int fa, Tensor inp, Tensor out, int quant_level, bool "
       "cast_bf2half) -> ()");
-  custom_ar.impl("qr_all_reduce", torch::kCUDA, &qr_all_reduce);
+  m.impl("qr_all_reduce", torch::kCUDA, &qr_all_reduce);
 
-  custom_ar.def("init_custom_qr", &init_custom_qr);
-  custom_ar.def("qr_destroy", &qr_destroy);
+  m.def("init_custom_qr", &init_custom_qr);
+  m.def("qr_destroy", &qr_destroy);
 
-  custom_ar.def("qr_get_handle", &qr_get_handle);
+  m.def("qr_get_handle", &qr_get_handle);
 
-  custom_ar.def("qr_open_handles(int _fa, Tensor[](b!) handles) -> ()");
-  custom_ar.impl("qr_open_handles", torch::kCPU, &qr_open_handles);
+  m.def("qr_open_handles(int _fa, Tensor[](b!) handles) -> ()");
+  m.impl("qr_open_handles", torch::kCPU, &qr_open_handles);
 
   // Max input size in bytes
-  custom_ar.def("qr_max_size", &qr_max_size);
+  m.def("qr_max_size", &qr_max_size);
 #endif
 
   /*
